atmospherex_activity_base/requirements.txt
apscheduler==3.6.3
coverage==5.2.1
fastapi==0.61.1
httpretty==1.0.2
pydantic==1.6.1
pytest==6.0.1
pytest-cov==2.10.1
pytest-env==0.6.2
pyyaml==5.3.1
requests==2.24.0
starlette==0.13.6
uvicorn==0.11.8